Add new app for legal issues
