+++ commons/templates/footer.html
@@ -28,11 +28,11 @@
       <div class="footer-column">
         <h2 class="subtitle"><i class="fas fa-gavel"></i> Legal</h2>
         <ul class="footer-column-items">
-          <li><a href="{% url 'events:coc' %}">Código de conducta</a></li>
-          <li><a href="{% url 'events:legal' %}">Aviso legal</a></li>
-          <li><a href="{% url 'events:privacy' %}">Política de privacidad</a></li>
-          <li><a href="{% url 'events:purchase_terms' %}">Condiciones de compra</a></li>
-          <li><a href="{% url 'events:cookie_policy' %}">Política de cookies</a></li>
+          <li><a href="{% url 'legal:legal_notice' %}">Aviso legal</a></li>
+          <li><a href="{% url 'legal:coc' %}">Código de conducta</a></li>
+          <li><a href="{% url 'legal:privacy_policy' %}">Política de privacidad</a></li>
+          <li><a href="{% url 'legal:purchase_terms' %}">Condiciones de compra</a></li>
+          <li><a href="{% url 'legal:cookie_policy' %}">Política de cookies</a></li>
         </ul>
       </div>
     </div>
