+++ events/templates/events/buy-article.html
@@ -70,12 +70,12 @@
       <div class="terms">
         <label class="checkbox">
           <input type="checkbox">
-          Acepto la <a href="{% url 'events:privacy' %}">Política de Privacidad</a>.
+          Acepto la <a href="{% url 'legal:privacy_policy' %}">Política de Privacidad</a>.
         </label>
 
         <label class="checkbox">
           <input type="checkbox">
-          Acepto las <a href="{% url 'events:purchase_terms' %}">Condiciones generales de compra</a>.
+          Acepto las <a href="{% url 'legal:purchase_terms' %}">Condiciones generales de compra</a>.
         </label>
       </div>
