+++ legal/static/legal/css/main.scss
@@ -1,0 +1,1 @@
+@import 'commons/static/commons/css/base';
