+++ legal/templates/legal/base.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %}about
+
+{% load utils %}
+
+{% block style %}
+  <link rel="stylesheet" href="{{ assets|get_asset_key:'legal/custom.min.css' }}">
+{% endblock style %}
