+++ legal/templates/legal/coc-en.html
@@ -1,4 +1,4 @@
-{% extends "events/base.html" %}
+{% extends "legal/base.html" %}
 {% load utils %}
 
 {% block title %}Code of Conduct - {{ block.super }}{% endblock %}
