+++ legal/templates/legal/coc-es.html
@@ -1,4 +1,4 @@
-{% extends "events/base.html" %}
+{% extends "legal/base.html" %}
 {% load utils %}
 
 {% block title %}Código de conducta - {{ block.super }}{% endblock %}
