+++ legal/templates/legal/cookie-policy.html
@@ -1,4 +1,4 @@
-{% extends "events/base.html" %}
+{% extends "legal/base.html" %}
 {% load utils %}
 
 {% block title %}Política de cookies - {{ block.super }}{% endblock %}
