+++ legal/templates/legal/privacy-policy.html
@@ -1,4 +1,4 @@
-{% extends "events/base.html" %}
+{% extends "legal/base.html" %}
 {% load utils %}
 
 {% block title %}Política de privacidad - {{ block.super }}{% endblock %}
