+++ legal/templates/legal/purchase-terms.html
@@ -1,4 +1,4 @@
-{% extends "events/base.html" %}
+{% extends "legal/base.html" %}
 {% load utils %}
 
 {% block title %}Condiciones generales de compra - {{ block.super }}{% endblock %}
